% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_builder.R
\name{build_all_assets}
\alias{build_all_assets}
\title{Build all site assets}
\usage{
build_all_assets(
  vendor_dir = "vendor/paws",
  build_dir = "build/mkdocs/docs",
  template_file = "build/mkdocs.orig.yml",
  output_yaml = "build/mkdocs/mkdocs.yml",
  alias_file = "build/aws_service_alias.yml",
  paws_dir = "vendor/paws/cran"
)
}
\arguments{
\item{vendor_dir}{Character string path to vendor directory.
Default is "vendor/paws"}

\item{build_dir}{Character string path to build output directory.
Default is "build/mkdocs/docs"}

\item{template_file}{Character string path to the original mkdocs template.
Default is "build/mkdocs.orig.yml"}

\item{output_yaml}{Character string path to the output mkdocs.yml file.
Default is "build/mkdocs/mkdocs.yml"}

\item{alias_file}{Character string path to AWS service alias YAML file.
Default is "build/aws_service_alias.yml"}

\item{paws_dir}{Character string path to paws directory.
Default is "vendor/paws/cran"}
}
\value{
NULL (invisibly)
}
\description{
Main wrapper function that orchestrates the entire site asset building process.
This includes copying assets, editing examples, and building the site YAML.
}
