% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd2md.R
\name{build_rd_to_md}
\alias{build_rd_to_md}
\title{Build Rd documentation and convert to Markdown}
\usage{
build_rd_to_md(
  dir = "vendor/paws/paws/man",
  md_dir = "build/mkdocs/docs/docs",
  long_doc = "vendor/paws/paws",
  cores = detectCores()
)
}
\arguments{
\item{dir}{Character string path to the directory containing Rd files.
Default is "vendor/paws/paws/man"}

\item{md_dir}{Character string path to output Markdown directory.
Default is "build/mkdocs/docs/docs"}

\item{long_doc}{Character string path to the package for building Rd docs.
Default is "vendor/paws/paws"}

\item{cores}{Integer number of cores to use for parallel processing.
Default is parallel::detectCores()}
}
\value{
NULL (invisibly)
}
\description{
Main function that orchestrates the entire Rd to Markdown conversion process.
This function:
1. Creates necessary directories
2. Builds Rd documentation
3. Collects all Rd files to process
4. Sets up parallel processing with mirai
5. Converts all Rd files to Markdown in parallel
}
