% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{check_pyktest}
\alias{check_pyktest}
\title{Check if \code{pyktest} Python package is available}
\usage{
check_pyktest(warn = TRUE)
}
\arguments{
\item{warn}{boolean, if TRUE (default), warn user about check result.}
}
\value{
boolean value indicating if the \code{pyktest} Python package is available
on the system.
}
\description{
Check if \code{pyktest} Python package is available
}
