% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{check_python}
\alias{check_python}
\title{Check if Python is available on the system and if its version is sufficient.}
\usage{
check_python(version = "3.5", warn = TRUE)
}
\arguments{
\item{version}{character string, Python required version (e.g. \code{"3.5"},
\verb{3.10.2}).}

\item{warn}{boolean, if TRUE (default), warn user about check result.}
}
\value{
boolean value indicating if the required Python version is available
on the system.
}
\description{
Check if Python is available on the system and if its version is sufficient.
}
\keyword{internal}
