% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearch_parsers.R
\name{chomp_hits}
\alias{chomp_hits}
\title{Hits to data.tables}
\usage{
chomp_hits(hits_json = NULL, keep_nested_data_cols = FALSE)
}
\arguments{
\item{hits_json}{A character vector. If its length is greater than 1, its elements will be pasted 
together. This can contain a JSON returned from a \code{search} query in Elasticsearch, or
a filepath or URL pointing at one. Most commonly, this JSON will be the direct result of 
running \code{\link[elastic]{Search}} with \code{raw = TRUE}.}

\item{keep_nested_data_cols}{a boolean (default FALSE); whether to keep columns that are nested
arrays in the original JSON. A warning will be given if these columns are deleted.}
}
\description{
A generic function for converting Elasticsearch docs into R data.tables. It
uses \code{\link[jsonlite]{fromJSON}} with \code{flatten = TRUE} to convert a
JSON into an R data.frame, and formats it into a data.table.
}
\examples{
\dontrun{
# Get 100 purchase records

# Get result from Elasticsearch
result <- uptasticsearch::es_search(es_host = "http://es.custdb.mycompany.com:9200"
                                    , es_index = "ticket_sales"
                                    , max_hits = 100)

# Parse into a data.table
resultDT <- chomp_hits(hits_json = result)
}
}
