% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{download_packs_and_deps}
\alias{download_packs_and_deps}
\title{Download a package and its deps for webR}
\usage{
download_packs_and_deps(
  pk_to_install,
  path_to_installation = "./webr_packages"
)
}
\arguments{
\item{pk_to_install}{what package do you want to install}

\item{path_to_installation}{Where should it be downloaded}
}
\description{
Download a package and its deps for webR
}
\examples{
if (FALSE){
 download_packs_and_deps("golem", tempdir())
}
}
