% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_desc.R
\name{download_packs_and_deps_from_desc}
\alias{download_packs_and_deps_from_desc}
\title{Download packages and deps from DESCRIPTION}
\usage{
download_packs_and_deps_from_desc(
  description,
  path_to_installation = "./webr_packages"
)
}
\arguments{
\item{description}{path to DESCRIPTION}

\item{path_to_installation}{passed to \code{download_pak_and_deps()}}
}
\description{
Download packages and deps from DESCRIPTION
}
