% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted-SVC_mle.R
\name{fitted.SVC_mle}
\alias{fitted.SVC_mle}
\title{Extact Model Fitted Values}
\usage{
\method{fitted}{SVC_mle}(object, ...)
}
\arguments{
\item{object}{\code{\link{SVC_mle}} object}

\item{...}{further arguments}
}
\value{
Data frame, fitted values to given data, i.e., the SVC as well as the response and their locations
}
\description{
Method to extract the fitted values from an \code{\link{SVC_mle}} object. This is only possible if \code{save.fitted} was set to \code{TRUE} in the control of the function call
}
\author{
Jakob Dambon
}
