% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kclass.R
\name{Fuller_k}
\alias{Fuller_k}
\title{Fuller K-class kappa}
\usage{
Fuller_k(alpha, A, X, Y, A_inc = NULL)
}
\arguments{
\item{alpha}{The Fuller hyperparameter  (numeric)}

\item{A}{exogenous variables (vector/matrix/data.frame)}

\item{X}{included endogenous regressors (vector/matrix/data.frame)}

\item{Y}{target variable (vector/matrix/data.frame)}

\item{A_inc}{included exogenous regressors (vector/matrix/data.frame)}
}
\value{
the Fuller estimator with hyperparameter alpha
}
\description{
`Fuller_k` compute the Fuller K-class kappa parameter for a given fuller hyperparameter, i.e., Fuller(alpha) estimate = Kclass(Fuller_k(alpha,...),...).
}
\details{
The Fuller(alpha=1) estimator is approximately unbiased and Fuller(alpha=4) approximately yields the minimal MSE estimator among all Fuller estimators.
}
