% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearch_eda_funs.R
\name{get_counts}
\alias{get_counts}
\title{Examine the distribution of distinct values for a field in Elasticsearch}
\usage{
get_counts(field, es_host, es_index, start_date = "now-1w",
  end_date = "now", use_na = "always", max_terms = 1000, time_field)
}
\arguments{
\item{field}{A valid field in whatever Elasticsearch index you are querying}

\item{es_host}{A string identifying an Elasticsearch host. This should be of the form 
[transfer_protocol][hostname]:[port]. For example, 'http://myindex.thing.com:9200'.}

\item{es_index}{The name of an Elasticsearch index to be queried.}

\item{start_date}{A character date of the form "yyyy-mm-dd", indicating the earliest
date from which to show documents. Set to NULL if you are looking at an index
for which limiting results by time is irrelevant. Default is "now-1w".}

\item{end_date}{A character date of the form "yyyy-mm-dd", indicating the most recent
date from which to show documents. Default is 10 "now". Set to NULL if you are 
looking at an index for which limiting results by time is irrelevant.}

\item{use_na}{Argument to control handling of missing values in the result. Options are
"always" to give a row in the table for NAs even if there are none, "ifany" to
include a count of missing values if there are any. If neither of these is passed
in, NAs will not be included in the result.}

\item{max_terms}{What is the maximum number of unique terms to return? Many production
Elasticsearch deployments limit this to a small number by default.}

\item{time_field}{Name of the date/time field in the target index that you want to filter on.}
}
\description{
For a given field, return a data table with its unique values
}
\examples{
\dontrun{
# Count number of customers by payment method
recoDT <- get_counts(field = "pmt_method"
                    , es_host = "http://es.custdb.mycompany.com:9200"
                    , es_index = "ticket_sales"
                    , start_date = "now-2w"
                    , end_date = "now")
}
}
