% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{get_examples}
\alias{get_examples}
\title{List the code example nav entries}
\usage{
get_examples(dir)
}
\arguments{
\item{dir}{Directory of example files (copied from vendor \verb{examples/}).}
}
\value{
A list of \code{"Display Name: examples/file"} nav entries.
}
\description{
List the code example nav entries
}
