% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester.R
\encoding{UTF-8}
\name{get_proj}
\alias{get_proj}
\title{Kernel projections}
\usage{
get_proj(kt, contrib = FALSE, t_max = NULL)
}
\arguments{
\item{kt}{a \code{Ktest} object, output of \code{\link[=ktest_init]{ktest_init()}} function.}

\item{contrib}{logical. If FALSE (default), a list of discriminant
axis projections per directions for each condition in the sample is
returned. If TRUE, a list of contributions to the discriminant
axis projections per directions for each condition in the sample is
returned.}

\item{t_max}{integer, the maximum truncation value to consider, i.e. the
maximum dimension for the kernel embedding projections. Default is NULL and
the truncation value will be the number of observations in the data.}
}
\value{
a list containing, for each sample condition, a projection matrix
or projection contribution matrix stored in a \code{data.frame} of dimension
the number of observations in the conditions × the number of dimensions up
to the truncation.
}
\description{
Get kernel projections for each condition with respect to the preset
truncation parameter for kernel FDA (kfda) test.
}
\details{
See Ozier-Lafontaine et al (2024) for more details.
}
\examples{
# WRITE ME
}
\references{
Ozier-Lafontaine A., Fourneaux C., Durif G., Arsenteva P., Vallot C.,
Gandrillon O., Gonin-Giraud S., Michel B., Picard F. (2024).
Kernel-Based Testing for Single-Cell Differential Analysis. Preprint.
\doi{10.48550/arXiv.2307.08509};
\href{https://arxiv.org/abs/2307.08509}{arXiv.2307.08509};
\href{https://hal.science/hal-04214858}{hal-04214858}.
}
