% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester.R
\name{get_pvalue}
\alias{get_pvalue}
\title{Get p-values table with respect to the preset truncation parameter.}
\usage{
get_pvalue(kt, contrib = False, log = False, verbose = 0)
}
\arguments{
\item{kt}{output of \code{\link[=ktest]{ktest()}} function.}

\item{contrib}{boolean, if \code{TRUE}, returns the p-value associated to
each principal component individually, otherwise returns the p-value
associated to the \code{kfda} statistic.}

\item{log}{boolean, if \code{TRUE}, returns the log p-values.}

\item{verbose}{integer (\code{1} or \code{0}) or logical (\code{TRUE} or \code{FALSE}),
enable/disable verbosity.}
}
\value{
FIXME
}
\description{
Get p-values table with respect to the preset truncation parameter.
}
\examples{
# WRITE ME
}
