% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester.R
\name{get_pvalues}
\alias{get_pvalues}
\title{Kernel test p-values}
\usage{
get_pvalues(kt, stat = "kfda", permutation = FALSE, t_max = NULL)
}
\arguments{
\item{kt}{a \code{Ktest} object, output of \code{\link[=ktest_init]{ktest_init()}} function.}

\item{stat}{character string
The test statistic to use, can be either \code{'kfda'} (default) or \code{'mmd'}.}

\item{permutation}{logical
If \code{FALSE} (default), the asymptotic approach is applied to compute
p-values. If \code{TRUE}, a permutation test is performed. If \code{stat = "mmd"}
then permutation based p-value computation is always used and
\code{permutation} value is ignored.}

\item{t_max}{integer, the maximum truncation value to consider, i.e. the
maximum dimension for the kernel embedding projections. Default is NULL and
the truncation value will be the number of observations in the data.}
}
\value{
a vector of p-values for each dimension up to the truncation value
or \code{NULL} value if corresponding \code{stat} was not computed.
}
\description{
Get p-values table with respect to the preset truncation parameter.
}
\details{
See Ozier-Lafontaine et al (2024) for more details.
}
\examples{
\dontrun{
# data loading
tmp <- load_example_data()
# gene expression data table (344 cells and 83 genes)
data_tab <- tmp$data_tab
# metadata table with sampling conditions (for the 344 cells)
metadata_tab <- tmp$metadata_tab

# create Ktest object
kt_1 = ktest_init(
    data = data_tab, metadata = metadata_tab, 
    sample_names = c('0H', '48HREV')
)
# run test
test(kt_1)
# get p-values
get_pvalues(kt_1)
}
}
\references{
Ozier-Lafontaine A., Fourneaux C., Durif G., Arsenteva P., Vallot C.,
Gandrillon O., Gonin-Giraud S., Michel B., Picard F. (2024).
Kernel-Based Testing for Single-Cell Differential Analysis. Preprint.
\doi{10.48550/arXiv.2307.08509};
\href{https://arxiv.org/abs/2307.08509}{arXiv.2307.08509};
\href{https://hal.science/hal-04214858}{hal-04214858}.
}
\seealso{
\code{\link[=ktest_init]{ktest_init()}}, \code{\link[=test]{test()}}, \code{\link[=get_statistics]{get_statistics()}},
\code{\link[=get_proj]{get_proj()}}
}
