% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester.R
\name{get_statistics}
\alias{get_statistics}
\title{Kernel test statistic}
\usage{
get_statistics(kt, stat = "kfda", contrib = FALSE, t_max = NULL)
}
\arguments{
\item{kt}{a \code{Ktest} object, output of \code{\link[=ktest_init]{ktest_init()}} function.}

\item{stat}{character string
The test statistic to use, can be either \code{'kfda'} (default) or \code{'mmd'}.}

\item{contrib}{logical, works only with \code{stat = 'kfda'}, if TRUE,
returns the vector of statistic contributions to each direction,
otherwise returns the cumulative statistic values with regards to the
direction. Ignored if \code{stat = 'mmd'}.}

\item{t_max}{integer, the maximum truncation value to consider, i.e. the
maximum dimension for the kernel embedding projections. Default is NULL and
the truncation value will be the number of observations in the data.}
}
\value{
a vector of statistic values for each dimension up to
the truncation or \code{NULL} value if corresponding \code{stat} was not computed.
}
\description{
Get statistic table with respect to the preset truncation parameter.
}
\details{
See Ozier-Lafontaine et al (2024) for more details.
}
\examples{
# WRITE ME
}
\references{
Ozier-Lafontaine A., Fourneaux C., Durif G., Arsenteva P., Vallot C.,
Gandrillon O., Gonin-Giraud S., Michel B., Picard F. (2024).
Kernel-Based Testing for Single-Cell Differential Analysis. Preprint.
\doi{10.48550/arXiv.2307.08509};
\href{https://arxiv.org/abs/2307.08509}{arXiv.2307.08509};
\href{https://hal.science/hal-04214858}{hal-04214858}.
}
\seealso{
\code{\link[=ktest_init]{ktest_init()}}, \code{\link[=test]{test()}}, \code{\link[=get_pvalues]{get_pvalues()}},
\code{\link[=get_proj]{get_proj()}}
}
