% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jgd-package.R
\docType{package}
\name{jgd-package}
\alias{jgd-package}
\title{jgd: JSON Graphics Device}
\description{
A lightweight (C-based, zero dependency) R graphics device that serializes
plotting operations into JSON and streams them to an external renderer.
See \code{\link[=jgd]{jgd()}} to open the device and \code{\link{jgd_spec}} for the wire protocol
specification.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/grantmcdermott/jgd}
  \item Report bugs at \url{https://github.com/grantmcdermott/jgd/issues}
}

}
\author{
\strong{Maintainer}: Grant McDermott \email{contact@grantmcdermott.com}

Authors:
\itemize{
  \item Grant McDermott \email{contact@grantmcdermott.com}
  \item Tatsuya Shima \email{ts1s1andn@gmail.com}
}

Other contributors:
\itemize{
  \item Dave Gamble (cJSON library in src/cjson/) [copyright holder]
  \item cJSON contributors (cJSON library in src/cjson/) [copyright holder]
}

}
\keyword{internal}
