% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kclass.R
\name{Kclass_lambda}
\alias{Kclass_lambda}
\title{K-class estimator (lambda penalty parametrization)}
\usage{
Kclass_lambda(lambda, A, Z, Y)
}
\arguments{
\item{lambda}{K-class penalty parameter (numeric)}

\item{A}{exogenous variables (vector/matrix/data.frame)}

\item{Z}{included exogenous and endogenous regressors (vector/matrix/data.frame)}

\item{Y}{target variable (vector/matrix/data.frame)}
}
\value{
the K-class estimator with penalty parameter lambda
}
\description{
K-class estimator (lambda penalty parametrization)
}
