% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktest-package.R
\docType{package}
\name{ktest-package}
\alias{ktest-package}
\alias{ktest}
\title{ktest}
\description{
Kernel-based testing
}
\details{
The \code{ktest} package implements kernel-based statistical testing, such as
maximal mean discrepancy test (MMD) and a test based on kernel Fisher
Discriminant Analysis (kFDA). It can be used for differential expression
analysis in transcriptomics data for instance.

See Ozier-Lafontaine et al (2024) for more details.
}
\references{
Ozier-Lafontaine A., Fourneaux C., Durif G., Arsenteva P., Vallot C.,
Gandrillon O., Gonin-Giraud S., Michel B., Picard F. (2024).
Kernel-Based Testing for Single-Cell Differential Analysis. Preprint.
\doi{10.48550/arXiv.2307.08509};
\href{https://arxiv.org/abs/2307.08509}{arXiv.2307.08509};
\href{https://hal.science/hal-04214858}{hal-04214858}.
}
\seealso{
\code{\link[=ktest_init]{ktest_init()}}, \code{\link[=test]{test()}}
}
\author{
\itemize{
\item \href{https://www.researchgate.net/profile/Anthony-Ozier-Lafontaine}{Anthony Ozier-Lafontaine}
\item \href{https://parsenteva.github.io/}{Polina Arsenteva}
\item \href{https://gdurif.perso.math.cnrs.fr/}{Ghislain Durif}
}
}
