% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester.R
\name{ktest}
\alias{ktest}
\title{Generate a Ktest object and compute specified comparison}
\usage{
ktest(
  data,
  metadata,
  condition,
  data_name = "data",
  samples = "all",
  var_metadata = NULL,
  stat = "kfda",
  nystrom = FALSE,
  permutation = FALSE,
  n_permutations = 500,
  seed_permutation = 0,
  test_params = NULL,
  center_by = NULL,
  marked_obs_to_ignore = NULL,
  kernel = NULL,
  verbose = 0
)
}
\arguments{
\item{data}{\code{data.frame} or equivalent containing the data with
observation in rows and variables/features in columns.}

\item{metadata}{\code{data.frame} or equivalent containing the metadata with
observation in rows and variables/features in columns.}

\item{condition}{: character string, column name of the metadata
\code{data.frame} containing the labels to test}

\item{data_name}{character string, name of the data (default = \code{"data"},
examples: \code{"counts"}, \code{"normalized"})}

\item{samples}{character string or vector/list of character string
(default = \code{"all"}):
\itemize{
\item character string: category of condition to compare to others as
\itemize{
\item \code{"all"}: test the two categories contained in the column condition
(does not work for more than two yet)
\item \code{"one vs all"}
}
\item vector/list of character string: the two categories to compare to
each other.
}}

\item{var_metadata}{\code{data.frame} or equivalent containing metainformation
about the variables, index must correspond to \code{data} columns
(default = \code{NULL}).}

\item{stat}{character string among \code{"kfda"}, \code{"mmd"} (default = \code{"kfda"}).
The test statistic to use. If \code{stat == "mmd"}, then
\code{permutation} is set to \code{TRUE}.}

\item{nystrom}{boolean (default = \code{FALSE}). If \code{TRUE}, the \code{test_params}
input argument is set to \code{TRUE} with default parameters.}

\item{permutation}{boolean, indicating whether to compute permutation or
asymptotic p-values(default = \code{FALSE}). Set to \code{TRUE} if \code{stat == "mmd"}.}

\item{n_permutations}{integer (default = \code{500}), number of permutations needed
to compute the permutation p-value. Ignored if \code{permutation} is \code{FALSE}.}

\item{seed_permutation}{integer (default = \code{0}), seed for the first
permutation. The other seeds are obtained by incrementation.}

\item{test_params}{object specifying the test to execute,
output of the function \code{get_test_params()}.}

\item{center_by}{character string (default = \code{NULL}),
column of the metadata containing a categorical variable to regress}

\item{marked_obs_to_ignore}{character string (default = \code{NULL}),
column of the metadata containing boolean values that are
\code{TRUE} if the observation should be ignored from the analysis
and \code{FALSE} otherwise.}

\item{kernel}{: object specifying the kernel function to use
output of the function \code{init_kernel_params()}.}

\item{verbose}{integer (\code{1} or \code{0}) or logical (\code{TRUE} or \code{FALSE}),
enable/disable verbosity.}
}
\value{
a \code{Ktest} object.
}
\description{
Generate a Ktest object and compute specified comparison
}
\examples{
# WRITE ME
}
