% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_example_data}
\alias{load_example_data}
\title{Load example dataset shipped with the package}
\usage{
load_example_data()
}
\value{
a list containing two data tables (\code{data.frame}):
\itemize{
\item \code{data_tab} with the gene expression measurements (in columns) for the
different cells (in rows);
\item \code{metadata_tab} with the corresponding condition labels.
}
}
\description{
Dataset containing single cell gene expression data and metadata
(i.e. corresponding experimental conditions).
}
\details{
This dataset originates from a study that investigated the molecular
mechanisms underlying cell differentiation and reversion, by measuring cell
transcriptomes at four time points: undifferentiated T2EC maintained in a
self-renewal medium (condition \code{"0H"}), then put in a
differentiation-inducing medium for 24h (condition \code{"24H"}). This population
was then split into a first population maintained in the same medium for
another 24h to achieve differentiation (condition \code{"48HDIFF"}), and the
second population was put back in the self-renewal medium to investigate
potential reversion (condition \code{"48HREV"}). Cell transcriptomes were
measured using scRT-qPCR on 83 genes selected to be involved in the
differentiation process.

See Zreika et al (2022) and Ozier-Lafontaine et al (2024) for more details.

The example dataset contains the samples for the condition \code{"0H"} and
\code{"48HREV"}.

See \url{https://github.com/LMJL-Alea/ktest/tree/main/tutorials/v5_data} for
the entire dataset.
}
\examples{
# data loading
tmp <- load_example_data()
# gene expression data table (344 cells and 83 genes)
data_tab <- tmp$data_tab
# metadata table with sampling conditions (for the 344 cells)
metadata_tab <- tmp$metadata_tab
}
\references{
Zreika S., Fourneaux C., Vallin E., Modolo L., Seraphin R., Moussy A.,
Ventre E., Bouvier M., Ozier-Lafontaine A., Bonnaffoux A., Picard F.,
Gandrillon O., Gonin-Giraud S. (2022 Jul 6).
Evidence for close molecular proximity between reverting and
undifferentiated cells. BMC Biol. 20(1):155.
\href{https://dx.doi.org/10.1186/s12915-022-01363-7}{doi:10.1186/s12915-022-01363-7};
\href{https://pubmed.ncbi.nlm.nih.gov/35794592/}{PMID: 35794592};
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC9258043/}{PMCID: PMC9258043};
\href{https://hal.science/hal-04134084v1}{hal-04134084v1}.

Ozier-Lafontaine A., Fourneaux C., Durif G., Arsenteva P., Vallot C.,
Gandrillon O., Gonin-Giraud S., Michel B., Picard F. (2024).
Kernel-Based Testing for Single-Cell Differential Analysis. Preprint.
\doi{10.48550/arXiv.2307.08509};
\href{https://arxiv.org/abs/2307.08509}{arXiv.2307.08509};
\href{https://hal.science/hal-04214858}{hal-04214858}.
}
