% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_info}
\alias{log_info}
\title{Write a timestamped INFO log line}
\usage{
log_info(msg)
}
\arguments{
\item{msg}{Character scalar message to log.}
}
\value{
\code{msg}, invisibly.
}
\description{
Write a timestamped INFO log line
}
