% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester.R
\name{multivariate_test}
\alias{multivariate_test}
\title{Compute multivariate testing}
\usage{
multivariate_test(
  kt,
  stat = NULL,
  permutation = NULL,
  n_permutations = 500,
  seed_permutation = 0,
  n_jobs_permutation = 1,
  keep_permutation_statistics = FALSE,
  verbose = 0
)
}
\arguments{
\item{kt}{output of \code{\link[=ktest]{ktest()}} function.}

\item{stat}{character string among \code{"kfda"}, \code{"mmd"} (default = \code{"kfda"}).
The test statistic to use. If \code{stat == "mmd"}, then
\code{permutation} is set to \code{TRUE}.}

\item{permutation}{boolean, indicating whether to compute permutation or
asymptotic p-values(default = \code{FALSE}). Set to \code{TRUE} if \code{stat == "mmd"}.}

\item{n_permutations}{integer (default = \code{500}), number of permutations needed
to compute the permutation p-value. Ignored if \code{permutation} is \code{FALSE}.}

\item{seed_permutation}{integer (default = \code{0}), seed for the first
permutation. The other seeds are obtained by incrementation.}

\item{verbose}{integer (\code{1} or \code{0}) or logical (\code{TRUE} or \code{FALSE}),
enable/disable verbosity.}
}
\value{
No return.
}
\description{
Compute multivariate testing
}
\examples{
# WRITE ME
}
