% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlocs.R
\name{nlocs}
\alias{nlocs}
\title{Extract Number of Unique Locations}
\usage{
nlocs(object)
}
\arguments{
\item{object}{\code{\link{SVC_mle}} object}
}
\value{
integer with the number of unique locations
}
\description{
Function to extract the number of unique locations in the data 
set used in an MLE of the \code{\link{SVC_mle}} object.
}
\author{
Jakob Dambon
}
