% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs-SVC_mle.R
\name{nobs.SVC_mle}
\alias{nobs.SVC_mle}
\title{Extract Number of Observations}
\usage{
\method{nobs}{SVC_mle}(object, ...)
}
\arguments{
\item{object}{\code{\link{SVC_mle}} object}

\item{...}{further arguments}
}
\value{
an integer of number of observations
}
\description{
Method to extract the number of observations used in MLE for an \code{\link{SVC_mle}} object.
}
\author{
Jakob Dambon
}
