% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.site.builder-package.R
\docType{package}
\name{paws.site.builder-package}
\alias{paws.site.builder}
\alias{paws.site.builder-package}
\title{paws.site.builder: Build Tools for Paws Documentation Site}
\description{
Provides functions for building the Paws R SDK documentation site. This package converts R documentation files (.Rd) to Markdown format and manages site assets, examples, and navigation structure for the MkDocs-based documentation site.
}
\author{
\strong{Maintainer}: DyfanJones \email{dyfan.r.jones@gmail.com}

}
\keyword{internal}
