% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\docType{package}
\name{pkg-package}
\alias{pkg}
\alias{pkg-package}
\title{pkg: Dev Package}
\description{
A package to develop the GitLab Actions workflow for ‘box’ without pushing lots of silly, broken commits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/klmr/r-pkg}
}

}
\author{
\strong{Maintainer}: Konrad Rudolph \email{konrad.rudolph@gmail.com} (\href{https://orcid.org/0000-0002-9866-7051}{ORCID})

}
\keyword{internal}
