% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester.R
\name{print_multivariate_test_results}
\alias{print_multivariate_test_results}
\title{Print a summary of the multivariate test results.}
\usage{
print_multivariate_test_results(kt, long = FALSE, t = NULL, ts = c(1, 5, 10))
}
\arguments{
\item{kt}{output of \code{\link[=ktest]{ktest()}} function.}

\item{long}{boolean (default = \code{FALSE}). If \code{TRUE}, the print is more
detailed.}

\item{t}{integer (default = \code{10}), if \code{long == FALSE}, value for truncation
parameter to be considered.}

\item{ts}{vector of integer (default = \code{c(1,5,10)}), if \code{long == FALSE},
values for truncation parameter to be considered.}
}
\value{
No return.
}
\description{
Print a summary of the multivariate test results.
}
\examples{
# WRITE ME
}
