% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pyktest}
\alias{pyktest}
\title{Global reference to main ktest Python module}
\format{
An object of class \code{python.builtin.module}
}
\value{
the \code{pyktest} Python module
}
\description{
The \code{pyktest} module object is an internal reference to the ktest
Python package that is used under the hood by \code{ktest} R package.
}
\keyword{internal}
