% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_envname}
\alias{random_envname}
\title{Helper function to generate random python environment name}
\usage{
random_envname(prefix = "", length = 10)
}
\arguments{
\item{prefix}{character string, prefix to add to the environment name.}

\item{length}{integer, number of random character in the environment name.}
}
\value{
character string, name of the python environment.
}
\description{
Helper function to generate random python environment name
}
\examples{
random_envname(prefix = "test", length = 10)
}
\keyword{internal}
