% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_rd.R
\name{render_rd_usage}
\alias{render_rd_usage}
\title{Render a \verb{\\usage} node into an \verb{## Usage} Markdown section}
\usage{
render_rd_usage(node)
}
\arguments{
\item{node}{A \verb{\\usage} Rd node.}
}
\value{
Character vector of Markdown lines.
}
\description{
Render a \verb{\\usage} node into an \verb{## Usage} Markdown section
}
