% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_rd.R
\name{render_rd_value}
\alias{render_rd_value}
\title{Render a \verb{\\value} node into an \verb{## Value} Markdown section}
\usage{
render_rd_value(node, alias_index = NULL)
}
\arguments{
\item{node}{A \verb{\\value} Rd node.}

\item{alias_index}{See \code{\link[=render_client_rd]{render_client_rd()}}.}
}
\value{
Character vector of Markdown lines.
}
\description{
Render a \verb{\\value} node into an \verb{## Value} Markdown section
}
