% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utils.R
\name{restore_html_definition_lists}
\alias{restore_html_definition_lists}
\title{Format Arguments section with indentation}
\usage{
restore_html_definition_lists(lines)
}
\arguments{
\item{lines}{Character vector of markdown lines}
}
\value{
Modified character vector with indented argument descriptions
}
\description{
Format Arguments section with indentation
}
