% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearch_eda_funs.R
\name{retrieve_mapping}
\alias{retrieve_mapping}
\title{Retrieve the mapping definitions for an index or index/type in Elasticsearch}
\usage{
retrieve_mapping(es_host, es_indexes = "_all", es_types = NULL)
}
\arguments{
\item{es_host}{A string identifying an Elasticsearch host. This should be of
the form \code{[transfer_protocol][hostname]:[port]}. For example,
\code{'http://myindex.thing.com:9200'}.}

\item{es_indexes}{A character vector that contains the names of indexes for
which to get mappings. Default, is \code{'_all'}, which means
get the mapping for all indexes}

\item{es_types}{A character vector that contains the names of types for which
to get mappings. Default is \code{NULL}, which means get the
mapping for all types in the chosen indexes}
}
\value{
A data.table containing four columns: index, type, field, and datatype
}
\description{
For one or multiple index or index/type, return a data table with
             field names and types.
}
\examples{
\dontrun{
# get the mapping for all types in the ticket_sales index
mappingDT <- retrieve_mapping(es_host = "http://es.custdb.mycompany.com:9200"
                              , es_indexes = "ticket_sales")
}
}
