% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rktest-package.R
\docType{package}
\name{rktest-package}
\alias{rktest}
\alias{rktest-package}
\title{rktest: Kernel Based Statistical Testing}
\description{
Write Me
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AnthoOzier/ktest}
}

}
\author{
\strong{Maintainer}: Ghislain Durif \email{gd.dev@libertymail.net} (\href{https://orcid.org/0000-0003-2567-1401}{ORCID})

Authors:
\itemize{
  \item Anthony Ozier-Lafontaine \email{anthony.ozier-lafontaine@ec-nantes.fr} (\href{https://orcid.org/0000-0003-3232-0503}{ORCID})
}

}
\keyword{internal}
