% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester.R
\name{test}
\alias{test}
\title{Compute multivariate testing}
\usage{
test(kt, stat = "kfda", permutation = FALSE, n_permutations = 500, verbose = 1)
}
\arguments{
\item{kt}{a \code{Ktest} object, output of \code{\link[=ktest_init]{ktest_init()}} function.}

\item{stat}{character string
The test statistic to use, can be either \code{'kfda'} (default) or \code{'mmd'}.}

\item{permutation}{logical
If \code{FALSE} (default), the asymptotic approach is applied to compute
p-values. If \code{TRUE}, a permutation test is performed. If \code{stat = "mmd"}
then permutation based p-value computation is always used and
\code{permutation} value is ignored.}

\item{n_permutations}{integer
Number of permutations performed. The default is 500.}

\item{verbose}{integer
The higher the verbosity, the more messages keeping track of
computations. The default is 1.
- < 1: no messages,
- 1: progress bar with computation time,
- 2: warnings are printed once,
- 3: warnings are printed every time they appear.}
}
\value{
No return. Input Ktest object \code{kt} is updated with test results.
}
\description{
Performs either the MMD or the kFDA test to compare the two samples of
the considered dataset. Stores the results in the corresponding attribute
depending on the statistic and on the p-value approach.
}
\details{
See 'ktest' package vignette for more details and usage examples. You can
also refer to Ozier-Lafontaine et al (2024) for more details about the
methods.
}
\examples{
\dontrun{
# data loading
tmp <- load_example_data()
# gene expression data table (344 cells and 83 genes)
data_tab <- tmp$data_tab
# metadata table with sampling conditions (for the 344 cells)
metadata_tab <- tmp$metadata_tab

# create Ktest object
kt_1 = ktest_init(
    data = data_tab, metadata = metadata_tab, 
    sample_names = c('0H', '48HREV')
)
# run test
test(kt_1)
# print results
print(kt_1)
}
}
\references{
Ozier-Lafontaine A., Fourneaux C., Durif G., Arsenteva P., Vallot C.,
Gandrillon O., Gonin-Giraud S., Michel B., Picard F. (2024).
Kernel-Based Testing for Single-Cell Differential Analysis. Preprint.
\doi{10.48550/arXiv.2307.08509};
\href{https://arxiv.org/abs/2307.08509}{arXiv.2307.08509};
\href{https://hal.science/hal-04214858}{hal-04214858}.
}
\seealso{
\code{\link[=ktest_init]{ktest_init()}}, \code{\link[=get_pvalues]{get_pvalues()}},
\code{\link[=get_statistics]{get_statistics()}}, \code{\link[=get_proj]{get_proj()}}
}
