% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_r_code.R
\name{wrap_r_examples}
\alias{wrap_r_examples}
\title{Wrap the "Examples" section in a code fence}
\usage{
wrap_r_examples(lines)
}
\arguments{
\item{lines}{Character vector of Markdown lines.}
}
\value{
Character vector of Markdown lines.
}
\description{
Wrap the "Examples" section in a code fence
}
