% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_r_code.R
\name{wrap_r_request_syntax}
\alias{wrap_r_request_syntax}
\title{Wrap the "Request syntax" section in a code fence}
\usage{
wrap_r_request_syntax(lines)
}
\arguments{
\item{lines}{Character vector of Markdown lines.}
}
\value{
Character vector of Markdown lines.
}
\description{
Wrap the "Request syntax" section in a code fence
}
