% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_block.R
\name{ai4bayescode_installed_blocks}
\alias{ai4bayescode_installed_blocks}
\title{Contributed blocks already installed in the user library (like `installed.packages()`)}
\usage{
ai4bayescode_installed_blocks()
}
\value{
Character vector of installed block names.
}
\description{
Contributed blocks already installed in the user library (like `installed.packages()`)
}
