% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utils.R
\name{build_long_rd}
\alias{build_long_rd}
\title{Build Rd documentation for long documentation packages}
\usage{
build_long_rd(long_doc = "vendor/paws/paws")
}
\arguments{
\item{long_doc}{Character string path to the package directory}
}
\value{
NULL (invisibly)
}
\description{
Build Rd documentation for long documentation packages
}
