% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_ktest}
\alias{check_ktest}
\title{Check if \code{ktest} package is ready to run.}
\usage{
check_ktest(verbose = TRUE)
}
\arguments{
\item{verbose}{boolean, if TRUE (default), inform user about check result.}
}
\value{
boolean value indicating if the \code{ktest} package is ready.
}
\description{
In practice, check if \code{ktest} Python package is installed (which should be
done by the \code{\link[=install_ktest]{install_ktest()}} function).
}
\examples{
\dontrun{
check_ktest()
}
}
\seealso{
\code{\link[=install_ktest]{install_ktest()}}
}
