% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_python}
\alias{check_python}
\title{Check Python availability}
\usage{
check_python(version = "3.12", verbose = TRUE)
}
\arguments{
\item{version}{character string, Python required version (e.g. \code{"3.5"},
\code{"3.10.2"}).}

\item{verbose}{boolean, if TRUE (default), inform user about check result.}
}
\value{
boolean value indicating if the required Python version is available
on the system.
}
\description{
Check if Python is available on the system and if its version is sufficient
(which should be automatically managed by \code{reticulate}).
}
\details{
This function is just a wrapper around \code{\link[reticulate:py_available]{reticulate::py_available()}} and
\code{\link[reticulate:py_version]{reticulate::py_version()}} functions.
}
\examples{
\dontrun{
check_python()
}
}
