% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{convert_name}
\alias{convert_name}
\title{Convert file names to title case}
\usage{
convert_name(file_names)
}
\arguments{
\item{file_names}{Character vector of file names}
}
\value{
Character vector of converted names
}
\description{
Convert file names to title case
}
