% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{edit_r_examples}
\alias{edit_r_examples}
\title{Edit R example files}
\usage{
edit_r_examples(dir = "build/mkdocs/docs/examples")
}
\arguments{
\item{dir}{Character string path to examples directory.
Default is "build/mkdocs/docs/examples"}
}
\value{
NULL (invisibly)
}
\description{
Converts R example files to Markdown format by wrapping them in code fences
and changing the file extension from .R to .md.
}
