% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{edit_readme}
\alias{edit_readme}
\title{Edit README for site}
\usage{
edit_readme(file = "build/mkdocs/docs/README.md")
}
\arguments{
\item{file}{Character string path to README.md file.
Default is "build/mkdocs/docs/README.md"}
}
\value{
NULL (invisibly)
}
\description{
Edits the README.md file to fix logo images, documentation links,
and example links for the site.
}
