% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utils.R
\name{find_and_replace}
\alias{find_and_replace}
\title{Find and replace operator references in HTML strings}
\usage{
find_and_replace(string, operator)
}
\arguments{
\item{string}{Character string to search}

\item{operator}{Character string operator name}
}
\value{
Modified string with links added
}
\description{
Find and replace operator references in HTML strings
}
