% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_hierarchy.R
\name{get_developer_guide}
\alias{get_developer_guide}
\title{Get developer guide files}
\usage{
get_developer_guide(dir = "build/mkdocs/docs/developer_guide")
}
\arguments{
\item{dir}{Character string path to developer_guide directory.
Default is "build/mkdocs/docs/developer_guide"}
}
\value{
Named list of developer guide files
}
\description{
Retrieves and formats developer guide documentation files.
}
