% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_hierarchy.R
\name{get_examples}
\alias{get_examples}
\title{Get example files}
\usage{
get_examples(dir = "build/mkdocs/docs/examples")
}
\arguments{
\item{dir}{Character string path to examples directory.
Default is "build/mkdocs/docs/examples"}
}
\value{
Named list of example files
}
\description{
Retrieves and formats example files.
}
