% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_hierarchy.R
\name{get_version}
\alias{get_version}
\title{Get package version}
\usage{
get_version(dir = "vendor/paws/cran/paws/DESCRIPTION")
}
\arguments{
\item{dir}{Character string path to DESCRIPTION file.
Default is "vendor/paws/cran/paws/DESCRIPTION"}
}
\value{
Character string version number
}
\description{
Extracts the version number from the DESCRIPTION file.
}
