% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_utils.R
\name{html_table_to_list}
\alias{html_table_to_list}
\title{Convert HTML tables to definition lists}
\usage{
html_table_to_list(lines)
}
\arguments{
\item{lines}{Character vector of HTML lines}
}
\value{
Modified character vector with tables converted to definition lists
}
\description{
Convert HTML tables to definition lists
}
