% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{install_ktest}
\alias{install_ktest}
\title{Install Python dependency}
\usage{
install_ktest(method = "auto", conda = "auto", ...)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}

\item{...}{Additional arguments passed to \code{\link[reticulate:conda_install]{conda_install()}}
or \code{\link[reticulate:virtualenv_install]{virtualenv_install()}}.}
}
\description{
The \code{ktest} R package relies on the \code{ktest} Python package
under the hood, thanks to the
\href{https://rstudio.github.io/reticulate/}{\code{reticulate}} package.
The \code{ktest} Python package should be installed for R \code{ktest} to run.
This function install Python \code{ktest} using the \code{reticulate} framework.
}
\details{
Python should be available on your system to install and use
Python \code{ktest}. You can run \code{\link[reticulate:py_discover_config]{reticulate::py_discover_config()}} to check
which version of Python will be used by \code{reticulate}.

To avoid messing with your system or user Python system environment, we
highly recommend using a Python virtual environment or a Conda environment
to install Python \code{ktest}.
See \href{https://rstudio.github.io/reticulate/articles/versions.html}{using Python with \code{reticulate}}
for more details.
}
\examples{
\dontrun{
# create Python environement
reticulate::virtualenv_create("ktest")
# activate Python environment
reticulate::use_virtualenv(virtualenv = "ktest", required = TRUE)
# check version of Python
reticulate::py_discover_config()
# install pykest
install_ktest(method = "virtualenv", envname = "ktest")
}
}
