% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jgd-package.R
\docType{package}
\name{jgd-package}
\alias{jgd-package}
\title{jgd: JSON Graphics Device}
\description{
A graphics device that translates R plotting operations into JSON and streams them over a local connection to an external display application. The device acts as a pure recorder with no rendering dependencies; all rendering occurs in that application (e.g. a 'VS Code' extension or a web browser). Official display applications are available from the project homepage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/grantmcdermott/jgd}
  \item Report bugs at \url{https://github.com/grantmcdermott/jgd/issues}
}

}
\author{
\strong{Maintainer}: Grant McDermott \email{contact@grantmcdermott.com}

Authors:
\itemize{
  \item Tatsuya Shima \email{ts1s1andn@gmail.com}
}

Other contributors:
\itemize{
  \item Dave Gamble (cJSON library in src/cjson/) [copyright holder]
  \item cJSON contributors (cJSON library in src/cjson/) [copyright holder]
}

}
\keyword{internal}
