% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jgd-package.R
\docType{package}
\name{jgd-package}
\alias{jgd-package}
\title{jgd: JSON Graphics Device}
\description{
A graphics device that records R plotting operations and streams them as JSON over a Unix domain socket to an external renderer. The device acts as a pure recorder with no rendering dependencies -- all rendering occurs in the client (e.g. a VS Code webview or any NDJSON-capable frontend).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/grantmcdermott/jgd}
  \item Report bugs at \url{https://github.com/grantmcdermott/jgd/issues}
}

}
\author{
\strong{Maintainer}: Grant McDermott \email{contact@grantmcdermott.com}

Authors:
\itemize{
  \item Tatsuya Shima \email{ts1s1andn@gmail.com}
}

}
\keyword{internal}
