% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_info}
\alias{log_info}
\title{Simple logger function}
\usage{
log_info(msg, ...)
}
\arguments{
\item{msg}{Character string containing the log message}
}
\value{
NULL (invisibly)
}
\description{
Simple logger function
}
