% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_hierarchy.R
\name{reference_index}
\alias{reference_index}
\title{Create reference index page}
\usage{
reference_index(
  paws_dir = "vendor/paws/cran",
  alias_file = "build/aws_service_alias.yml",
  output_file = "build/mkdocs/docs/docs/reference_index.md"
)
}
\arguments{
\item{paws_dir}{Character string path to paws directory.
Default is "vendor/paws/cran"}

\item{alias_file}{Character string path to AWS service alias YAML file.
Default is "build/aws_service_alias.yml"}

\item{output_file}{Character string path to output reference index file.
Default is "build/mkdocs/docs/docs/reference_index.md"}
}
\value{
Character string path to the reference index file
}
\description{
Generates a reference index page listing all available AWS services
organized by package.
}
