% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_wrappers.R
\name{wrap_r_examples}
\alias{wrap_r_examples}
\title{Wrap R examples section in code fences}
\usage{
wrap_r_examples(lines)
}
\arguments{
\item{lines}{Character vector of markdown lines}
}
\value{
Modified character vector
}
\description{
Wrap R examples section in code fences
}
