% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_wrappers.R
\name{wrap_r_request_syntax}
\alias{wrap_r_request_syntax}
\title{Wrap R request syntax section in code fences}
\usage{
wrap_r_request_syntax(lines)
}
\arguments{
\item{lines}{Character vector of markdown lines}
}
\value{
Modified character vector
}
\description{
Wrap R request syntax section in code fences
}
