% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_wrappers.R
\name{wrap_r_service_syntax}
\alias{wrap_r_service_syntax}
\title{Wrap R service syntax section in code fences}
\usage{
wrap_r_service_syntax(lines)
}
\arguments{
\item{lines}{Character vector of markdown lines}
}
\value{
Modified character vector
}
\description{
Wrap R service syntax section in code fences
}
