% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_wrappers.R
\name{wrap_r_usage}
\alias{wrap_r_usage}
\title{Wrap R usage section in code fences}
\usage{
wrap_r_usage(lines)
}
\arguments{
\item{lines}{Character vector of markdown lines}
}
\value{
Modified character vector
}
\description{
Wrap R usage section in code fences
}
